duh i cannot cli git
