+++ lib/deps.ts
@@ -1,2 +1,10 @@
 export { default as StrictEventEmitter } from "https://raw.githubusercontent.com/bterlson/strict-event-emitter-types/master/src/index.ts";
 export { default as EventEmitter } from "https://deno.land/std@v0.50.0/node/events.ts";
+export {
+    connectWebSocket,
+    isWebSocketCloseEvent,
+    WebSocket,
+    WebSocketCloseEvent,
+} from "https://deno.land/std@v0.50.0/ws/mod.ts";
+export { equal } from "https://deno.land/std@v0.50.0/bytes/mod.ts";
+export { default as pako } from "https://raw.githubusercontent.com/Denocord/pako/master/mod.js";
