+++ mod.ts
@@ -1,5 +1,7 @@
-import { bus } from "./lib/Client.ts";
+import { bus } from "./lib/client.ts";
 import WebsocketShard from "./lib/ws/websocket_shard.ts";
+
+export const VERSION = "0.0.1";
 
 export function onDebug(debugHandler: (msg: string) => void) {
   bus.on("debug", debugHandler);
@@ -8,7 +10,7 @@ export function onDebug(debugHandler: (msg: string) => void) {
 export function onError(errorHandler: (error: Error) => void) {
   bus.on("error", errorHandler);
 }
+const ws = WebsocketShard.get();
+export const login = ws.login.bind(ws);
 
-export const { login } = WebsocketShard.get();
-
-export { config } from "./lib/Client.ts";
+export { config } from "./lib/client.ts";
