+++ rest.ts
@@ -1,0 +1,2 @@
+import RequestHandler from "./lib/rest/request_handler.ts";
+export default RequestHandler.get();
