--- src/ClientOptions.ts
@@ -1,7 +1,0 @@
-import { Gateway } from "./@types/denocord.ts";
-
-export interface ClientOptions {
-  compress?: boolean;
-  compressStream?: boolean;
-  intents?: Gateway.GatewayIntents;
-}
