--- src/deps.ts
@@ -1,10 +1,0 @@
-export {
-  connectWebSocket,
-  isWebSocketCloseEvent,
-  WebSocket,
-  WebSocketCloseEvent,
-} from "https://deno.land/std@v0.50.0/ws/mod.ts";
-export * as log from "https://deno.land/std@v0.50.0/log/mod.ts";
-export { equal } from "https://deno.land/std@v0.50.0/bytes/mod.ts";
-export { default as pako } from "https://raw.githubusercontent.com/Denocord/pako/master/mod.js";
-export { default as EventEmitter } from "https://deno.land/std@v0.50.0/node/events.ts";
