--- src/mod.ts
@@ -1,3 +1,0 @@
-export { default as WebsocketShard } from "./gateway/WebsocketShard.ts";
-export { default as Client } from "./Client.ts";
-export * from "./@types/denocord.ts";
